Refactor login/register/change password to auth
