+++ src/routes/index.route.ts
@@ -7,13 +7,15 @@ import emailRoute from './email.route';
 import categoryRoute from './category.route';
 import uploadRoute from './upload.route';
 import fileRoute from './file.route';
+import authRoute from './auth.route';
 
 const router = express.Router();
 
+router.use('/auth', authRoute);
 router.use('/admin', adminRoute);
 router.use('/category', categoryRoute);
-router.use('/file', fileRoute);
 router.use('/email', emailRoute);
+router.use('/file', fileRoute);
 router.use('/mentorship', mentorship);
 router.use('/upload', uploadRoute);
 router.use('/user', userRoute);
