+++ src/routes/schema/user.schema.ts
@@ -2,30 +2,10 @@ import joi, { string } from 'joi';
 import { PRIVACY_PERMISSIONS_ENUM } from '../../constants/enum';
 
 export default {
-  login: joi.object({
-    email: joi.string().email().required(),
-    password: joi.string().required(),
-  }),
-
-  register: joi.object({
-    newUser: joi.object({
-      username: joi.string().required(),
-      email: joi.string().email().required(),
-      password: joi.string().required(),
-      confirmPassword: joi.string().required(),
-      isStudent: joi.boolean().required(),
-    }),
-  }),
-
   accountIdQ: joi.object({
     accountId: joi.string().required(),
   }),
 
-  changePassword: joi.object({
-    oldPassword: joi.string().required(),
-    newPassword: joi.string().required(),
-    confirmPassword: joi.string().required(),
-  }),
   createExperienceB: joi.object({
     experience: joi.object({
       role: joi.string().required(),
