+++ src/routes/user.route.ts
@@ -10,27 +10,6 @@ const router = express.Router();
 
 const schemaValidator = require('express-joi-validation').createValidator({});
 
-// ================================ USER AUTH ================================
-router.post(
-  '/login',
-  schemaValidator.body(user.login),
-  Utility.asyncHandler(UserController.login)
-);
-
-router.post(
-  '/register',
-  schemaValidator.body(user.register),
-  Utility.asyncHandler(UserController.register)
-);
-
-// authentication: check that req.user == accountId
-router.put(
-  '/change-password',
-  passport.authenticate('isAuthenticated', { session: false }),
-  schemaValidator.body(user.changePassword),
-  Utility.asyncHandler(UserController.changePassword)
-);
-
 // ================================ USER ================================
 
 // get user (student/sensei)
