+++ src/services/user.service.ts
@@ -1,114 +1,10 @@
 import { User } from '../models/User';
-import bcrypt from 'bcrypt';
 import { STATUS_ENUM, USER_TYPE_ENUM } from '../constants/enum';
 import { Admin } from '../models/Admin';
 import { ERRORS } from '../constants/errors';
 import { Op } from 'sequelize';
 import { Experience } from '../models/Experience';
 export default class UserService {
-  // ================================ USER AUTH ================================
-
-  public static async changePassword(
-    accountId: string,
-    userType: USER_TYPE_ENUM,
-    oldPassword: string,
-    newPassword: string,
-    confirmPassword: string
-  ) {
-    try {
-      const user = await this.findUserOrAdminById(accountId, userType);
-      const correctOldPassword = await bcrypt.compare(
-        oldPassword,
-        user.password
-      );
-
-      // throw error if old password is incorrect
-      if (!correctOldPassword) throw new Error('Old password is incorrect');
-
-      // throw error if newPassword != confirmPassword
-      if (newPassword != confirmPassword)
-        throw new Error('New password does not match');
-
-      if (newPassword == oldPassword)
-        throw new Error('New Password cannot be the same as the Old Password');
-      // change pass
-      const salt = await bcrypt.genSalt(10);
-      const hash = await bcrypt.hash(newPassword, salt);
-      user.password = hash;
-      user.save();
-    } catch (e) {
-      throw e;
-    }
-  }
-
-  public static async register(registerBody: {
-    username: string;
-    email: string;
-    password: string;
-    confirmPassword: string;
-    isStudent: boolean;
-  }): Promise<User> {
-    const {
-      username,
-      email,
-      password,
-      confirmPassword,
-      isStudent,
-    } = registerBody;
-
-    if (!username || !email || !password || !confirmPassword) {
-      throw new Error('Please enter all fields');
-    }
-
-    if (password != confirmPassword) {
-      throw new Error('Passwords do not match');
-    }
-
-    if (password.length < 8) {
-      throw new Error('Password must be at least 8 characters');
-    }
-    let user, newUser;
-
-    // check if user exist as a student or sensei
-    try {
-      user = await User.findOne({
-        where: {
-          [Op.or]: [{ email }, { username }],
-        },
-        paranoid: false,
-      });
-      if (isStudent) {
-        newUser = new User({
-          username,
-          email,
-          password,
-          userType: USER_TYPE_ENUM.STUDENT,
-        });
-      } else {
-        newUser = new User({
-          username,
-          email,
-          password,
-          userType: USER_TYPE_ENUM.SENSEI,
-        });
-      }
-
-      // if user exist, return error
-      if (user) {
-        throw new Error('Email/Username already exists');
-      }
-
-      // hash password
-      const salt = await bcrypt.genSalt(10);
-      const hash = await bcrypt.hash(newUser.password, salt);
-      newUser.password = hash;
-      newUser.save();
-      return newUser;
-    } catch (e) {
-      throw e;
-    }
-  }
-
   // ================================ USER ================================
 
   public static async deactivateUser(accountId: string): Promise<void> {
